requirements.txt
wheel>=0.36.2
requests==2.25.1
web3==5.19.0
bscscan-python==1.0.0
pyetherbalance==0.0.1
peewee==3.14.4
pytest==6.2.4
pyyaml==5.4.1